git-svn-id: https://svn.syssoft.uni-trier.de/svn/prisma/DeltaWorks/include@1834 87af6b1e-9b34-e211-997a-bc305bce1b9a

+++ general/system.cpp
@@ -287,7 +287,7 @@ namespace System
 	{
 		#if SYSTEM==WINDOWS
 			DWORD rs = 0;
-			BOOL rc = PeekNamedPipe(read_handle,target,bytes,&rs,NULL,NULL);
+			BOOL rc = PeekNamedPipe(read_handle,target,(DWORD)bytes,&rs,NULL,NULL);
 			if (!rc)
 				return 0;
 			return rs;
